libpeas: ignore GParameter deprecation
The API we use here relies on GParameter, and while it may be deprecated
from GLib, the API will have to live for some time in libpeas itself.

Instead, we can disable deprecations and rely on the GLib compiler
warnings for projects that are consuming these APIs.

Doing so also frees up a lot of our console output in builds to ensure
people see other deprecations that need to be adjusted.