build: hide private symbols with -fvisibility=hidden
This continues with the previous commit to hide non-public symbols by
default. It is slightly complicated by the fact that we need some symbol
access from the unit tests. Normally that would mean we just add a static
library and link against that, but we need shared access to globals so
that is not an option. We could possibly switch to creating a shared
library without -fvisibility=hidden.