CI: fix CCACHE_* env vars values
Previously, the basedir env var didn't expand
and the final value of the config looked like
${CI_PROJECT_DIR}/_ccache/ccache.conf as opposed to
an actuall path.

When gitlab parses the yml file, `CI_PROJECT_DIR` is a known
value that can be expanded, while our custom variable for
CCACHE_DIR isn't fully expanded due to pointing to a custom and
thus ignored env var.

Set the CCACHE_DIR to "$CI_PROJECT_DIR/_ccache" instead of
CCACHE_BASEDIR to workaround this issue and have it properly expand.