libpeas-gtk: fix gtk_menu_popup() deprecations
As libpeas-gtk only requires GTK 3.0, this attempts to fix the menu
positioning code in a backwards compatible way (at compile time). If we
have at least GTK 3.22, we can use the new popup APIs which can vastly
simplify this code and also fix menus displayed at monitor boundaries
on wayland.

The existing code was left mostly untouched to reduce the chances of
breaking things on older systems and also make it convenient to delete
once libpeas moves into the GTK 4.0 territory if/when that happens.