peas-object-module.c: Add fallback module search for Windows
When we specify the plugin module name via a plugin file,
g_module_build_path() will append a 'lib' prefix to the final filename
of the module DLL, which may cause the module load to fail because the
module DLL is not built with the 'lib' prefix for its filename, which
is often the case for Meson builds with Visual Studio and also likely
in other scenarios.

Fix this by adding a fallback check for Windows, that we look for the
module DLL without the 'lib' prefix by appending '.dll' to the module
name that we pass into g_module_build_path(), if we can't find the
module DLL with the 'lib' prefix.