meson.build: Improve lua search on Windows
On Windows, especially Visual Studio builds, lua does not come with a
pkg-config file, so we may need to look for the dependencies manually.

We first reverse the order of the search in such situations, by looking
for whether we have lua.exe or luajit.exe in our path, and then we look
for the lua headers and the lua library[1].  If they could be found, we
proceed to look for lua-lgi and build as we did before.

[1]: lua51.lib for luajit or lua5[3|2|1].lib for lua 5.[3|2|1].