build: Fix installation and some GTK tests
Don't specify the install directory for libpeas and libpeas-gtk, as we
need the DLLs to be in $(prefix)/bin and the .lib's/.so's to be in
$(libdir), which Meson sorts out for us automatically.

Also, just export the internal symbols for the tests when we build
libpeas-gtk, as GObject on Windows really doesn't like the idea of
mixing static .lib's with DLLs (i.e. a static libpeas-gtk mixing with a
DLL libpeas).