tests: Explicitly depend on the GIR and typelib
In Meson since 0.55.0, having the test executable depend on the GIR
dependency is not enough to ensure that uninstalled GIR and typelib
files are generated: we also need to have the test itself depend on
the GIR build target.

It is not clear to me whether this is a workaround for a Meson
regression (if what libpeas previously did was meant to work), or a
solution to a libpeas bug (if it was never meant to be guaranteed to
work). For more details see
<https://github.com/mesonbuild/meson/issues/7756>.

Resolves: https://gitlab.gnome.org/GNOME/libpeas/-/issues/40
Bug-Debian: https://bugs.debian.org/966951