engine: sort unresolved plugins at tail of queue
If we have a plugin we've added that we could not resolve all the
dependencies for then push it to the tail of the queue. That way, further
insertions of plugins are more likely to sort before the plugin.

Otherwise, we can push the info to the head as it has no dependencies.

This improves a situation in Builder where plugins that dependend on
plugins which also had dependencies could end up loading in the wrong
order.