engine: implement GListModel
This commit makes PeasEngine a GListModel containing PeasPluginInfo. Such
a design is useful for modern GTK development because it allows binding
the engine to various list widgetry.

Additionally, you may use a GtkFilterListModel or others to alter or even
map the PeasPluginInfo into other forms.

This also removes the G_TYPE_POINTER based plugin-list property as that
is no longer needed (as well as rather unsafe).