build: Improve Python dep search
Extend the 'python3_path' option,if set, to use a Python installation to
override the default one that is found via pkg-config (or from the
installation of Python that is used to invoke Meson), on all platforms
supported by libpeas.

This will make things easier for people who wish to build against
multiple installations of Python without rebuilding everything.