tests: Drop weak pointer line
My theory is that the extension's last ref is the toggle ref owned by
GJS, so the weak pointer triggers after the test is completed and the
scope is exited. At that point, &extension no longer exists because it
pointed into test_extension_gjs_instance_refcount()'s stack, and so it
smashes some other function's stack.

Removing this line fixes the crash.