tests: port to girepository-2.0
With GNOME/glib!4630 we can start using girepository-2.0 and still
integrate with the bindings in use. However, this still requires that the
bindings are using girepository-2.0 too.

It is expected that we can do that for GNOME 49 across the various modules
and this ensures we will have working unit tests as that work is completed.

GJS has patches ready but lua-gi still needs patching.